rewrote the choose() function to be more effecient
I rewrote the choose() function to work with percentage point
probabilities for transitions (as opposed to the hacky old way which
appened a new item for every transition to approximate it with a random
index). This sorts the probabilities and then walks through them in
descending order, comparing them to a randomly generated value between 0
and 1. For example

c = 0.23 // random choice

choice 1: 20%
choice 2: 40%
choice 3: 40%

if 0 < c < 0.20:
	pick choice 1
else if 0.20 < c < 0.60: // 0.20 + 0.40 = 0.60
	pick choice 2
else if 0.60 < c < 1.00: // 0.60 + 0.40 = 1.00
	pick choice 3
	
You can see how the probability of each choice maps to a certain "range"
of numbers between 0 and 1.